backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
netmiko==4.3.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
statsmodels==0.14.2
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
reportlab==4.1.0
websockets==12.0
nmap==0.7.1
ipaddress==1.0.23
asyncio-mqtt==0.16.2
prometheus-fastapi-instrumentator==7.0.0
